backend/requirements.txt
# FastAPI and web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# GraphQL
strawberry-graphql[fastapi]>=0.215.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# TimescaleDB (PostgreSQL extension - installed separately)
# Redis
redis>=5.0.0
hiredis>=2.2.3

# Workflow Engine (Temporal replaces Cadence)
temporalio>=1.8.0

# Data validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.26.0

# Data sources
yfinance>=0.2.0

# Scientific computing
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

# Geospatial
h3>=3.7.0

# Quantitative libraries
# quantlib-python>=1.17  # Optional - install separately if needed
pyportfolioopt>=1.5.0
cvxpy>=1.4.0

# Graph algorithms
networkx>=3.2.1

# Export formats
pyarrow>=14.0.0  # For Parquet
orjson>=3.9.10  # Fast JSON

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.26.0  # For async testing

# Code quality
ruff>=0.1.8
mypy>=1.7.0
types-redis>=4.6.0

# Environment variables
python-dotenv>=1.0.0

